migrate account roles data source to autogen docs
