+++ docs/data-sources/account_roles.md
@@ -1,18 +1,19 @@
 ---
-layout: "cloudflare"
-page_title: "Cloudflare: cloudflare_account_roles"
-description: Get information on a Cloudflare Account Roles.
+page_title: "cloudflare_account_roles Data Source - Cloudflare"
+subcategory: ""
+description: |-
+  Use this data source to lookup Account Roles https://api.cloudflare.com/#account-roles-properties.
 ---
 
-# cloudflare_account_roles
+# cloudflare_account_roles (Data Source)
 
-Use this data source to lookup [Account Roles][1].
+Use this data source to lookup [Account Roles](https://api.cloudflare.com/#account-roles-properties).
 
-## Example usage
+## Example Usage
 
-```hcl
+```terraform
 data "cloudflare_account_roles" "account_roles" {
-    account_id = var.cloudflare_account_id
+    account_id = "goo"
 }
 
 locals {
@@ -30,18 +31,25 @@ resource "cloudflare_account_member" "member" {
 }
 ```
 
-## Argument Reference
+<!-- schema generated by tfplugindocs -->
+## Schema
 
-- `account_id` - (Required) The account for which to list the roles.
+### Required
 
-## Attributes Reference
+- `account_id` (String) The account identifier to target for the resource.
 
-- `roles` - A list of roles object. See below for nested attributes.
+### Read-Only
 
-**roles**
+- `id` (String) The ID of this resource.
+- `roles` (List of Object) A list of roles object. (see [below for nested schema](#nestedatt--roles))
 
-- `id` - Role identifier tag
-- `name` - Role Name
-- `description` - Description of role's permissions
+<a id="nestedatt--roles"></a>
+### Nested Schema for `roles`
 
-[1]: https://api.cloudflare.com/#account-roles-properties
+Read-Only:
+
+- `description` (String)
+- `id` (String)
+- `name` (String)
+
+
