+++ internal/provider/schema_data_source_account_roles.go
@@ -1,0 +1,45 @@
+package provider
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+func dataSourceCloudflareAccountRoles() *schema.Resource {
+	return &schema.Resource{
+		ReadContext: dataSourceCloudflareAccountRolesRead,
+
+		Schema: map[string]*schema.Schema{
+			"account_id": {
+				Description: "The account identifier to target for the resource.",
+				Type:        schema.TypeString,
+				Required:    true,
+			},
+
+			"roles": {
+				Type:        schema.TypeList,
+				Computed:    true,
+				Description: "A list of roles object.",
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"id": {
+							Type:        schema.TypeString,
+							Optional:    true,
+							Description: "Role identifier tag.",
+						},
+						"name": {
+							Type:        schema.TypeString,
+							Optional:    true,
+							Description: "Role Name.",
+						},
+						"description": {
+							Type:        schema.TypeString,
+							Optional:    true,
+							Description: "Description of role's permissions.",
+						},
+					},
+				},
+			},
+		},
+		Description: "Use this data source to lookup [Account Roles](https://api.cloudflare.com/#account-roles-properties).",
+	}
+}
